# backend/src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "test Application API",
    "description": "test Application API"
  },
  "host": "localhost:3000",
  "basePath": "/api",
  "tags": [
    {
      "name": "Users",
      "description": "API for users in the system"
    }
  ],
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/users": {
      "post": {
        "tags": ["Users"],
        "summary": "Create new user in system",
        "description": "Create new user in system",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User that we want to create",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "New user is created",
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "get": {
        "tags": ["Users"],
        "summary": "Get all users in system",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Users"
            }
          }
        }
      },
      "put": {
        "tags": ["Users"],
        "summary": "Updates the user",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "User payload for update",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserViewModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": ["Users"],
        "summary": "Get user by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the user",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/UserViewModel"
            }
          }
        }
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Delete user by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the user",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "required": ["username"],
      "properties": {
        "userName": {
          "type": "string",
          "pattern": "^[a-zA-Z0-9]*$",
          "minLength": 3,
          "maxLength": 30
        }
      }
    },
    "UserViewModel": {
      "properties": {
        "id": {
          "type": "string",
          "examples": ["5f8abd12a57f4e103cb4d302"]
        },
        "username": {
          "type": "string",
          "examples": ["u001"]
        },
        "createdAt": {
          "type": "string",
          "examples": ["2020-10-17T09:44:50.543Z"]
        }
      }
    },
    "Users": {
      "type": "array",
      "$ref": "#/definitions/UserViewModel"
    }
  }
}